Fixed email validation and handler functions; unify string quotes
