+++ models/__init__.py
@@ -4,4 +4,4 @@ from .Phone import Phone
 from .Record import Record
 from .AddressBook import AddressBook
 from .Birthday import Birthday
-from .email_validator import Email
+from .email import Email
